Move README and all internal code to jabber.org
The only change to the code is the removal of most `require` forms,
and some Org-ification of the README. I have not deleted the elisp
files, in case we decide to change our tangling strategy down the line
and need the per-file contributor information and copyright notices.

The external notification libraries have not yet been moved to jabber.org.