Resolve reference to free variable ‘widget’.
The widget library has apparently changed.  The referenced
template (widget-color-complete - mentioned in comments) no longer
exists.  There is still support for ‘:complete-function’ in
‘widget-default-completions’.

But since it ‘:complete-function’ invokes the function argument with
’call-interactive’, the widget parameter cannot be passed.  This used
to work due to dynamic binding.  But with lexical binding this will
no longer work.

Better to use ‘:complete’ which is invoked (non-interactively) and is
passed a ‘widget’ parameter,