Tame MUC presence notifications.
This patch hides or applies a face to MUC presence notifications in
the MUC chat buffer.  To control its behavior, customize
’jabber-muc-decorate-presence-patterns’.  By default it does nothing.

’jabber-muc-decorate-presence-patterns’ is a list of pairs consisting
of a regular expression and a either a face or ‘nil’.  If a the
regular expression matches a presence notification, then either:

 - the specified face is applied to the notification message
 - or if the second value of the pair is nil, the notification is
   discarded

If no regular expression in the list of pairs matches the notification
message, the message is displayed unchanged.

For example, the customization:

    '(jabber-muc-decorate-presence-patterns
      '(("\\( enters the room ([^)]+)\\| has left the chatroom\\)$")
        ("." . jabber-muc-presence-dim)))

hides participant enter/leave notifications.  It also diminishes other
presence notification messages to make it easier to distinguish
between conversation and notifications.