Expand nick completions with c-a-p, not hippie.
‘completion-at-point’ provides a more flexible completion framework
than current mechanism tightly bound to ‘hippie-expand’.  The current
mechanism is, after all, labeled with a "One big hack" comment.

This patch replaces the original ‘hippie-expand’ integration with
completion based on ‘completion-at-point’.  This can be leveraged by,
for example, adding:

    (set (make-local-variable 'completion-styles)
          '(substring basic partial-completion emacs22))

to ‘jabber-chat-mode-hook’.  This would allow completing a nick name
by typing a substring of the nickname, then pressing TAB.  This could
be used for completing nicknames that start with characters which are
cumbersome to insert.