Resolve reference to free variable ‘xml-data’.
Vestige of dynamic binding.  ‘jabber-presence-only-chat-open-message’
depends on access to the input message XML payload.  This function is
not even bound by default.  The documentation suggests
‘jabber-presence-only-chat-open-message’ can be used as an alternative
value for `jabber-alert-presence-message-function'.

The alternative resolution to the free variable reference, passing
‘xml-data’ as a parameter, turned into a holy mess producing a diff
with 51 insertions and 35 deletions.  That was a mess.  Seems prudent
to simply dynamic bind that one variable explicitly; even if it is
unclean that anyone uses ‘jabber-presence-only-chat-open-message’.