Fix variable/function warnings part two of three.
Fix most of the remaining variations of function not known to be
defined or reference or assignment to free variable warnings.  Like
the last patch, this patch is aimed at eliminating
native compiler warning message.

Where the first patch only manipulated ‘require’ functions, this
change only adds ‘defvar’ and ‘define-function’ forms to resolve the
remaining cross-file references.

With this patch applied there are still a few remaining warnings about
functions and variables.  These remaining warnings are related to the
order of definition and reference in a single file.  Those warnings
will be resolved by the next commit.