Fix variable/function warnings part one of three.
Fix variations of function not known to be defined or reference or
assignment to free variable warnings.  These messages arise from the
native compiler, even on files that ‘byte-compile’ successfully.

This first patch simply adds or removes ‘require’ functions to intern
unresolved symbols.  In some cases, remove ‘require’ functions to
prevent circular dependencies.

In particular, jabber-core, jabber-chat, and jabber-muc are
significantly intertwined.  So importing those in the wrong
combination quickly leads to native compiler to throw errors related
to circular dependencies.

Refactoring ‘require’ functions leaves quite a few undeclared
references.  These will be fixed in the next two commits.

The circular dependencies argues for a dependency graph and possible
refactoring.  But this is not that.  Reducing the dependency graph to
a DAC, though ideal, would require significant refactoring.