Remove two function references.
Promoting symbol quotes to function quotes uncovers two problematic
function references.

The key binding in jabber-roster.el for `jabber-ft-send' references a
deprecated function.  Since file transfer has been deprecated, do not
expose it in the keymap.

A search of the project git log suggests that the function
`jabber-activity-remove-visible' has never been defined.  So there is
no reason to try to remove it from `window-configuration-change-hook'.

To make that reference even more obscure, the `remove-hook' is only
performed if running under XEmacs.  Perhaps it is time to stop
pretending to support XEmacs.