Fix jabber-pkg.el for MELPA
`package-build', the build tool used by Melpa, requires that the
`define-package' form is the first form in this file.  This package
is the only package that provides such a file but does not satisfy
that requirement.

I don't know whether Tromey's ELPA actually still requires that this
file first explicitly requires `package' (if it actually ever did),
and whether Jabber is still being published there.