Fix broken advice :after #’jabber-send-presence.
2e034c3185 broke the advice that jabber-mode-line-mode used to display
self and contact status in the mode-line.  It partially, but not
completely, converted the obsolete `defadvice' mechanism for defining
advice to the replacement `advice-add' mechanism.

But it did so globally, leaving the obsolete `ad-activate' in place.
The combination of the new `advice-add' and obsolete `ad-activate'
generates the error in issue #52.

This completes the conversion by replacing `ad-activate' with
`advice-add', invoking `advice-add' only when enabling
`jabber-mode-line-mode'.

"(elisp) Advising Named Functions" suggests that advice is not the
best mechanism for updating the mode line: "... advice should be
reserved for the cases where you cannot modify a function’s behavior
in any other way."

We will get to that.