Suppress inactive participants in nick completion.
In large rooms, with minimal or no nick prefix, completion via c-a-p
produces a list of nick alternatives large enough to fill the screen.
This obscures, rather than helps identify recent posters.

It is unusual to address or reference lurkers or others that have not
posted to the room.  So it is not useful to include them in the nick
completion list.